Evaluates brains based on fitness
Evaluates the bunnies' brains based on the fitness function, but fitness
function is currently only called(updated) when evaluate is called.

+++ CaptureTheCarrot/Assets/Scripts/INeuralNetwork.cs
@@ -31,7 +31,7 @@ namespace AssemblyCSharp
 				
 				double[] getWeights();
 				
-				int Evaluate();
+				float Evaluate();
 		}
 }
